feat(ui): menu simplifié + recherche + page Billets (catalogue)

+++ SCinema/FilmsController.cs
@@ -1,15 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using SCinema.Data;
 using SCinema.Models;
 
 namespace SCinema
 {
+    [Authorize(Roles = "Fournisseur")]
     public class FilmsController : Controller
     {
         private readonly ApplicationDbContext _context;
@@ -22,114 +21,75 @@ namespace SCinema
         // GET: Films
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Films.ToListAsync());
+            var films = await _context.Films
+                .OrderBy(f => f.Title)     // <- ton champ Title
+                .ToListAsync();
+            return View(films);
         }
 
         // GET: Films/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
-            var film = await _context.Films
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (film == null)
-            {
-                return NotFound();
-            }
+            var film = await _context.Films.FirstOrDefaultAsync(m => m.Id == id);
+            if (film == null) return NotFound();
 
             return View(film);
         }
 
         // GET: Films/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
+        public IActionResult Create() => View();
 
         // POST: Films/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Title,Synopsis,Genre,DureeMinutes,AffichereUrl,DateSortie")] Film film)
+        public async Task<IActionResult> Create([Bind("Title,Synopsis,Genre,DureeMinutes,AfficheUrl,DateSortie")] Film film)
         {
-            if (ModelState.IsValid)
-            {
-                _context.Add(film);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(film);
+            if (!ModelState.IsValid) return View(film);
+
+            _context.Add(film);
+            await _context.SaveChangesAsync();
+            return RedirectToAction(nameof(Index));
         }
 
         // GET: Films/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var film = await _context.Films.FindAsync(id);
-            if (film == null)
-            {
-                return NotFound();
-            }
+            if (film == null) return NotFound();
             return View(film);
         }
 
         // POST: Films/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Title,Synopsis,Genre,DureeMinutes,AffichereUrl,DateSortie")] Film film)
         {
-            if (id != film.Id)
-            {
-                return NotFound();
-            }
+            if (id != film.Id) return NotFound();
+            if (!ModelState.IsValid) return View(film);
 
-            if (ModelState.IsValid)
+            try
             {
-                try
-                {
-                    _context.Update(film);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!FilmExists(film.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
+                _context.Update(film);
+                await _context.SaveChangesAsync();
             }
-            return View(film);
+            catch (DbUpdateConcurrencyException)
+            {
+                if (!_context.Films.Any(e => e.Id == film.Id)) return NotFound();
+                throw;
+            }
+            return RedirectToAction(nameof(Index));
         }
 
         // GET: Films/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
-            var film = await _context.Films
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (film == null)
-            {
-                return NotFound();
-            }
+            var film = await _context.Films.FirstOrDefaultAsync(m => m.Id == id);
+            if (film == null) return NotFound();
 
             return View(film);
         }
@@ -140,18 +100,9 @@ namespace SCinema
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var film = await _context.Films.FindAsync(id);
-            if (film != null)
-            {
-                _context.Films.Remove(film);
-            }
-
+            if (film != null) _context.Films.Remove(film);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
-        }
-
-        private bool FilmExists(int id)
-        {
-            return _context.Films.Any(e => e.Id == id);
         }
     }
 }
