+++ SCinema/SallesController.cs
@@ -1,7 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -10,6 +10,7 @@ using SCinema.Models;
 
 namespace SCinema
 {
+    [Authorize(Roles = "Fournisseur")]
     public class SallesController : Controller
     {
         private readonly ApplicationDbContext _context;
@@ -22,25 +23,24 @@ namespace SCinema
         // GET: Salles
         public async Task<IActionResult> Index()
         {
-            var applicationDbContext = _context.Salles.Include(s => s.Fournisseur);
-            return View(await applicationDbContext.ToListAsync());
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier)!;
+            var list = await _context.Salles
+                .Where(s => s.FournisseurId == userId)
+                .OrderBy(s => s.NumeroSalle)
+                .ToListAsync();
+            return View(list);
         }
 
         // GET: Salles/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier)!;
+            if (id == null) return NotFound();
 
             var salle = await _context.Salles
-                .Include(s => s.Fournisseur)
+                .Where(s => s.FournisseurId == userId)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (salle == null)
-            {
-                return NotFound();
-            }
+            if (salle == null) return NotFound();
 
             return View(salle);
         }
@@ -48,95 +48,82 @@ namespace SCinema
         // GET: Salles/Create
         public IActionResult Create()
         {
-            ViewData["FournisseurId"] = new SelectList(_context.Users, "Id", "Id");
+            // pour ne pas casser la vue scaffold : une liste contenant uniquement l'utilisateur courant
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier)!;
+            ViewData["FournisseurId"] = new SelectList(new[] { new { Id = userId } }, "Id", "Id", userId);
             return View();
         }
 
         // POST: Salles/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,NumeroSalle,Capacite,FournisseurId")] Salle salle)
+        public async Task<IActionResult> Create([Bind("NumeroSalle,Capacite,FournisseurId")] Salle salle)
         {
-            if (ModelState.IsValid)
+            if (!ModelState.IsValid)
             {
-                _context.Add(salle);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                var userIdForm = User.FindFirstValue(ClaimTypes.NameIdentifier)!;
+                ViewData["FournisseurId"] = new SelectList(new[] { new { Id = userIdForm } }, "Id", "Id", userIdForm);
+                return View(salle);
             }
-            ViewData["FournisseurId"] = new SelectList(_context.Users, "Id", "Id", salle.FournisseurId);
-            return View(salle);
+
+            // on force le propriétaire
+            salle.FournisseurId = User.FindFirstValue(ClaimTypes.NameIdentifier)!;
+
+            _context.Add(salle);
+            await _context.SaveChangesAsync();
+            return RedirectToAction(nameof(Index));
         }
 
         // GET: Salles/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier)!;
+            if (id == null) return NotFound();
 
-            var salle = await _context.Salles.FindAsync(id);
-            if (salle == null)
-            {
-                return NotFound();
-            }
-            ViewData["FournisseurId"] = new SelectList(_context.Users, "Id", "Id", salle.FournisseurId);
+            var salle = await _context.Salles
+                .Where(s => s.FournisseurId == userId)
+                .FirstOrDefaultAsync(s => s.Id == id);
+            if (salle == null) return NotFound();
+
+            ViewData["FournisseurId"] = new SelectList(new[] { new { Id = userId } }, "Id", "Id", userId);
             return View(salle);
         }
 
         // POST: Salles/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,NumeroSalle,Capacite,FournisseurId")] Salle salle)
         {
-            if (id != salle.Id)
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier)!;
+            if (id != salle.Id) return NotFound();
+
+            // vérifier la propriété
+            var exists = await _context.Salles.AnyAsync(s => s.Id == id && s.FournisseurId == userId);
+            if (!exists) return NotFound();
+
+            if (!ModelState.IsValid)
             {
-                return NotFound();
+                ViewData["FournisseurId"] = new SelectList(new[] { new { Id = userId } }, "Id", "Id", userId);
+                return View(salle);
             }
 
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(salle);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!SalleExists(salle.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            ViewData["FournisseurId"] = new SelectList(_context.Users, "Id", "Id", salle.FournisseurId);
-            return View(salle);
+            salle.FournisseurId = userId; // on garde le proprio
+            _context.Update(salle);
+            await _context.SaveChangesAsync();
+
+            return RedirectToAction(nameof(Index));
         }
 
         // GET: Salles/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier)!;
+            if (id == null) return NotFound();
 
             var salle = await _context.Salles
-                .Include(s => s.Fournisseur)
+                .Where(s => s.FournisseurId == userId)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (salle == null)
-            {
-                return NotFound();
-            }
+            if (salle == null) return NotFound();
 
             return View(salle);
         }
@@ -146,19 +133,15 @@ namespace SCinema
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var salle = await _context.Salles.FindAsync(id);
-            if (salle != null)
-            {
-                _context.Salles.Remove(salle);
-            }
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier)!;
 
+            var salle = await _context.Salles
+                .Where(s => s.FournisseurId == userId)
+                .FirstOrDefaultAsync(s => s.Id == id);
+
+            if (salle != null) _context.Salles.Remove(salle);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
-        }
-
-        private bool SalleExists(int id)
-        {
-            return _context.Salles.Any(e => e.Id == id);
         }
     }
 }
